+++ lib/main.dart
@@ -1,9 +1,11 @@
+import 'package:firebase_core/firebase_core.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:package_info_plus/package_info_plus.dart';
 
-import 'config/constants.dart';
+import 'config/firebase_options_dev.dart' as dev;
+import 'config/firebase_options_prod.dart' as prod;
 import 'domain/entity/app_info.dart';
 import 'domain/repository/stamp_rally/stamp_rally_repository.dart';
 import 'domain/repository/user/user_repository.dart';
@@ -12,13 +14,18 @@ import 'infrastructure/firebase/stamp_rally/stamp_rally_repository.dart';
 import 'infrastructure/firebase/user/user_repository.dart';
 import 'presentation/app.dart';
 import 'util/assets/assets.gen.dart';
+import 'util/constants.dart';
 import 'util/provider_logger.dart';
 
 Future<void> main() async {
   WidgetsFlutterBinding.ensureInitialized();
 
   // Firebase の初期化
-  await initializeFirebaseApp();
+  await Firebase.initializeApp(
+    options: isProd
+        ? prod.DefaultFirebaseOptions.currentPlatform
+        : dev.DefaultFirebaseOptions.currentPlatform,
+  );
 
   // 画面の向きを固定
   await SystemChrome.setPreferredOrientations([
